# documentation/openapi.yaml
openapi: 3.0.1
info:
  title: Musicman API
  description: Описание всех эндпоинтов сервиса "Musicman"
  contact:
    name: Анастасия Сашина
    email: sashina@cs.vsu.ru
  version: 1.0.6
servers:
  - url: http://localhost:8080/api/v1
    description: Generated server url
security:
  - Bearer: []
paths:
...
  "/playlists/{id}":
    get:
      tags:
        - Playlist Controller
      summary: Выводит плейлист пользователя с песнями
      operationId: getPlaylistById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Возвращает плейлист
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PlaylistWithSongsDTO"
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Плейлист не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
    put:
      tags:
        - Playlist Controller
      summary: Обновление плейлиста
      operationId: updatePlaylist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdatePlaylistRequest"
        required: true
      responses:
        '200':
          description: Плейлист обновлен
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Плейлист не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
    delete:
      tags:
        - Playlist Controller
      summary: Удаление плейлиста
      operationId: deletePlaylist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Плейлист удален
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Плейлист не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
...
  "/playlists":
    get:
      tags:
        - Playlist Controller
      summary: Выводит все плейлисты пользователя
      operationId: getAllPlaylists
      parameters:
        - name: pageNumber
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 5
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Возвращает все плейлисты
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PlaylistPagesDTO"
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
    post:
      tags:
        - Playlist Controller
      summary: Создает новый плейлист
      operationId: createPlaylist
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreatePlaylistRequest"
        required: true
      responses:
        '201':
          description: Плейлист создан
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Плейлист не существует
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
  "/playlists/{id}/add-song/{songId}":
    post:
      tags:
        - Playlist Controller
      summary: Добавить песню в плейлист
      operationId: addSongToPlaylist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: songId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Песня добавлена
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Плейлист или песня не существуют
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '409':
          description: Песня уже есть в плейлисте
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
...
  "/playlists/{id}/exists/{songId}":
    get:
      tags:
        - Playlist Controller
      summary: Выводит информацию о существовании песни в плейлисте
      operationId: isSongInPlaylist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: songId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Возвращает результат
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SongExistsDTO"
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Плейлист или песня не существуют
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
...
  "/playlists/{id}/delete-song/{songId}":
    delete:
      tags:
        - Playlist Controller
      summary: Удаление песни из плейлиста
      operationId: deleteSongFromPlaylist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: songId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Песня удалена
          content:
            application/json: {}
        '400':
          description: Невалидные входные данные
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '404':
          description: Плейлист или песня не существуют
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
        '405':
          description: Песни нет в плейлисте
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExceptionDTO"
components:
  schemas:
    ExceptionDTO:
      type: object
      properties:
        message:
          type: string
...
    UpdatePlaylistRequest:
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 0
          type: string
        description:
          type: string
...
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
...
    CreatePlaylistRequest:
      required:
        - description
        - name
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 0
          type: string
        description:
          type: string
...
    GenreDTO:
      type: object
      properties:
        genreId:
          type: integer
          format: int64
        name:
          type: string
...
    PagingInfoDTO:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
    SingerDTO:
      type: object
      properties:
        singerId:
          type: integer
          format: int64
        fullName:
          type: string
        description:
          type: string
    SongDTO:
      type: object
      properties:
        songId:
          type: integer
          format: int64
        title:
          type: string
        duration:
          "$ref": "#/components/schemas/LocalTime"
        singer:
          "$ref": "#/components/schemas/SingerDTO"
        genres:
          uniqueItems: true
          type: array
          items:
            "$ref": "#/components/schemas/GenreDTO"
...
    PlaylistDTO:
      type: object
      properties:
        playlistId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    PlaylistPagesDTO:
      type: object
      properties:
        paging:
          "$ref": "#/components/schemas/PagingInfoDTO"
        data:
          type: array
          items:
            "$ref": "#/components/schemas/PlaylistDTO"
    PlaylistWithSongsDTO:
      type: object
      properties:
        playlistInfo:
          "$ref": "#/components/schemas/PlaylistDTO"
        songs:
          uniqueItems: true
          type: array
          items:
            "$ref": "#/components/schemas/SongDTO"
    SongExistsDTO:
      type: object
      properties:
        exists:
          type: boolean
...
  securitySchemes:
    Authorization:
      type: http
      scheme: Bearer
      bearerFormat: JWT